+++ deploy/complete/terraform/oke-autoscaler.tf
@@ -1,11 +1,11 @@
-# Copyright (c) 2021 Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2021, 2022, Oracle and/or its affiliates. All rights reserved.
 # Licensed under the Universal Permissive License v 1.0 as shown at http://oss.oracle.com/licenses/upl.
 # 
 
 locals {
-  cluster_autoscaler_supported_k8s_versions           = ["1.17", "1.18", "1.19"]
-  cluster_autoscaler_image_version                    = "2021.03"
-  cluster_autoscaler_image                            = "iad.ocir.io/oracle/oci-cluster-autoscaler:${local.k8s_major_minor_version}-${local.cluster_autoscaler_image_version}"
+  cluster_autoscaler_supported_k8s_versions           = { "1.18" = "1.18.3-4", "1.19" = "1.19.1-4", "1.20" = "1.20.0-4", "1.21" = "1.21.1-1" } # There's no API to get that list. Need to be updated manually
+  cluster_autoscaler_image_version                    = lookup(local.cluster_autoscaler_supported_k8s_versions, local.k8s_major_minor_version, reverse(values(local.cluster_autoscaler_supported_k8s_versions))[0])
+  cluster_autoscaler_image                            = "iad.ocir.io/oracle/oci-cluster-autoscaler:${local.cluster_autoscaler_image_version}"
   cluster_autoscaler_app_version                      = 4
   cluster_autoscaler_node_pool                        = var.create_new_oke_cluster ? oci_containerengine_node_pool.oke_node_pool[0].id : var.existent_oke_nodepool_id_for_autoscaler
   cluster_autoscaler_min_nodes                        = var.cluster_autoscaler_min_nodes
@@ -14,7 +14,7 @@ locals {
   cluster_autoscaler_scale_down_delay_after_add       = "10m"
   cluster_autoscaler_scale_down_unneeded_time         = "10m"
   cluster_autoscaler_unremovable_node_recheck_timeout = "5m"
-  cluster_autoscaler_enabled                          = contains(local.cluster_autoscaler_supported_k8s_versions, local.k8s_major_minor_version) ? var.cluster_autoscaler_enabled : false
+  cluster_autoscaler_enabled                          = contains(keys(local.cluster_autoscaler_supported_k8s_versions), local.k8s_major_minor_version) ? var.cluster_autoscaler_enabled : false
   k8s_major_minor_version                             = regex("\\d+(?:\\.(?:\\d+|x)(?:))", (var.k8s_version == "Latest") ? local.node_pool_k8s_latest_version : var.k8s_version)
 }
 resource "kubernetes_service_account" "cluster_autoscaler_sa" {
@@ -30,7 +30,7 @@ resource "kubernetes_service_account" "cluster_autoscaler_sa" {
   }
   automount_service_account_token = true # workaround to support ORM deprecated terraform providers
 
-  depends_on = [oci_containerengine_node_pool.oke_node_pool, local_file.kubeconfig]
+  depends_on = [oci_containerengine_node_pool.oke_node_pool]
 }
 resource "kubernetes_cluster_role" "cluster_autoscaler_cr" {
   count = local.cluster_autoscaler_enabled ? 1 : 0
@@ -111,7 +111,7 @@ resource "kubernetes_cluster_role" "cluster_autoscaler_cr" {
     verbs          = ["get", "update"]
   }
 
-  depends_on = [oci_containerengine_node_pool.oke_node_pool, local_file.kubeconfig]
+  depends_on = [oci_containerengine_node_pool.oke_node_pool]
 }
 resource "kubernetes_role" "cluster_autoscaler_role" {
   count = local.cluster_autoscaler_enabled ? 1 : 0
@@ -137,7 +137,7 @@ resource "kubernetes_role" "cluster_autoscaler_role" {
     verbs          = ["delete", "get", "update", "watch"]
   }
 
-  depends_on = [oci_containerengine_node_pool.oke_node_pool, local_file.kubeconfig]
+  depends_on = [oci_containerengine_node_pool.oke_node_pool]
 }
 resource "kubernetes_cluster_role_binding" "cluster_autoscaler_crb" {
   count = local.cluster_autoscaler_enabled ? 1 : 0
@@ -160,7 +160,7 @@ resource "kubernetes_cluster_role_binding" "cluster_autoscaler_crb" {
     namespace = "kube-system"
   }
 
-  depends_on = [oci_containerengine_node_pool.oke_node_pool, local_file.kubeconfig]
+  depends_on = [oci_containerengine_node_pool.oke_node_pool]
 }
 resource "kubernetes_role_binding" "cluster_autoscaler_rb" {
   count = local.cluster_autoscaler_enabled ? 1 : 0
@@ -185,7 +185,7 @@ resource "kubernetes_role_binding" "cluster_autoscaler_rb" {
     namespace = "kube-system"
   }
 
-  depends_on = [oci_containerengine_node_pool.oke_node_pool, local_file.kubeconfig]
+  depends_on = [oci_containerengine_node_pool.oke_node_pool]
 }
 resource "kubernetes_deployment" "cluster_autoscaler_deployment" {
   count = local.cluster_autoscaler_enabled ? 1 : 0
@@ -240,10 +240,16 @@ resource "kubernetes_deployment" "cluster_autoscaler_deployment" {
             "--v=${local.cluster_autoscaler_app_version}",
             "--stderrthreshold=info",
             "--cloud-provider=oci",
+            "--max-node-provision-time=${local.cluster_autoscaler_max_node_provision_time}",
             "--nodes=${local.cluster_autoscaler_min_nodes}:${local.cluster_autoscaler_max_nodes}:${local.cluster_autoscaler_node_pool}",
             "--scale-down-delay-after-add=${local.cluster_autoscaler_scale_down_delay_after_add}",
             "--scale-down-unneeded-time=${local.cluster_autoscaler_scale_down_unneeded_time}",
-            "--unremovable-node-recheck-timeout=${local.cluster_autoscaler_unremovable_node_recheck_timeout}"
+            "--unremovable-node-recheck-timeout=${local.cluster_autoscaler_unremovable_node_recheck_timeout}",
+            "--balance-similar-node-groups",
+            "--balancing-ignore-label=displayName",
+            "--balancing-ignore-label=hostname",
+            "--balancing-ignore-label=internal_addr",
+            "--balancing-ignore-label=oci.oraclecloud.com/fault-domain"
           ]
           image_pull_policy = "Always"
           env {
@@ -255,5 +261,5 @@ resource "kubernetes_deployment" "cluster_autoscaler_deployment" {
     }
   }
 
-  depends_on = [oci_containerengine_node_pool.oke_node_pool, local_file.kubeconfig]
+  depends_on = [oci_containerengine_node_pool.oke_node_pool]
 }